better organization for baseline files
