+++ front/scripts/baseline/mercury/utils/isPrimitive.ts
@@ -1,0 +1,10 @@
+module Mercury.Utils {
+	export function isPrimitive (val: any) {
+		var typeOf = typeof val;
+		return (val === null) ||
+				(typeOf === 'string') ||
+				(typeOf === 'number') ||
+				(typeOf === 'boolean') ||
+				(typeOf === 'undefined');
+	}
+}
