+++ front/scripts/main/app.ts
@@ -1,7 +1,6 @@
 /// <reference path="../../../typings/jquery/jquery.d.ts" />
 /// <reference path="../../../typings/ember/ember.d.ts" />
 /// <reference path="../../../typings/i18next/i18next.d.ts" />
-/// <reference path="../baseline/mercury.d.ts" />
 /// <reference path="../baseline/mercury.ts" />
 /// <reference path="../mercury/utils/track.ts" />
